♻️:  하이라이트 기능을 위한 highlightText 함수를 HighlightedText 컴포넌트로 리팩토링
- 하이라이트 계산 함수 및 하이라이트 태그 렌더링 함수 추출
